+++ src/Enums/ResponseFormat.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Busarm\PhpMini\Enums;
+
+/**
+ * PHP Mini Framework
+ *
+ * @copyright busarm.com
+ * @license https://github.com/Busarm/php-mini/blob/master/LICENSE (MIT License)
+ * @codeCoverageIgnore
+ */
+class ResponseFormat
+{
+    const JSON  =   'json';
+    const HTML  =   'html';
+    const XML   =   'xml';
+}