+++ src/Interfaces/RouteInterface.php
@@ -75,7 +75,7 @@ interface RouteInterface
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
      * @return static
      */
-    public static function get(string $path): RouteInterface;
+    public static function get(string $path): static;
 
     /**
      * Set HTTP POST routes
@@ -83,7 +83,7 @@ interface RouteInterface
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
      * @return static
      */
-    public static function post(string $path): RouteInterface;
+    public static function post(string $path): static;
 
     /**
      * Set HTTP PUT routes
@@ -91,7 +91,7 @@ interface RouteInterface
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
      * @return static
      */
-    public static function put(string $path): RouteInterface;
+    public static function put(string $path): static;
 
     /**
      * Set HTTP PATCH routes
@@ -99,7 +99,7 @@ interface RouteInterface
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
      * @return static
      */
-    public static function patch(string $path): RouteInterface;
+    public static function patch(string $path): static;
 
     /**
      * Set HTTP DELETE routes
@@ -107,5 +107,5 @@ interface RouteInterface
      * @param string $path HTTP path. e.g /home. See `Router::MATCHER_REGX` for list of parameters matching keywords
      * @return static
      */
-    public static function delete(string $path): RouteInterface;
+    public static function delete(string $path): static;
 }