+++ src/Middlewares/ResponseMiddleware.php
@@ -6,10 +6,7 @@ use Busarm\PhpMini\App;
 use Busarm\PhpMini\Dto\BaseDto;
 use Busarm\PhpMini\Dto\CollectionBaseDto;
 use Busarm\PhpMini\Interfaces\MiddlewareInterface;
-use Busarm\PhpMini\Interfaces\ResponseInterface;
-use Busarm\PhpMini\View;
-
-use function Busarm\PhpMini\Helpers\is_cli;
+use Busarm\PhpMini\Interfaces\ResponseHandlerInterface;
 
 /**
  * PHP Mini Framework
@@ -23,10 +20,8 @@ final class ResponseMiddleware implements MiddlewareInterface
     {
         $response = $next ? $next() : null;
         if ($response !== false) {
-            if ($response instanceof ResponseInterface) {
-                return is_cli() ? $response : $response->send($app->config->httpResponseFormat, $app->config->httpSendAndContinue);
-            } else if ($response instanceof View) {
-                return $response->send($app->config->httpSendAndContinue);
+            if ($response instanceof ResponseHandlerInterface) {
+                return $response->handle($app->config->httpSendAndContinue);
             } else if ($response instanceof CollectionBaseDto) {
                 return $app->sendHttpResponse(200, $response->toArray());
             } else if ($response instanceof BaseDto) {